+++ towerhttp/compress_gzip.go
@@ -13,20 +13,40 @@ type GzipCompression struct {
 	pool *sync.Pool
 }
 
+// NewGzipCompression creates a new GzipCompression.
+func NewGzipCompression() *GzipCompression {
+	return &GzipCompression{
+		pool: &sync.Pool{
+			New: func() interface{} {
+				return new(bytes.Buffer)
+			},
+		},
+	}
+}
+
 func (g GzipCompression) ContentEncoding() string {
 	return "gzip"
 }
 
-func (g GzipCompression) Compress(b []byte) ([]byte, error) {
+func (g GzipCompression) Compress(b []byte) ([]byte, bool, error) {
 	if len(b) <= 1500 {
-		return b, nil
+		return b, false, nil
 	}
 	buf := g.pool.Get().(*bytes.Buffer) //nolint
 	buf.Reset()
 	w, _ := gzip.NewWriterLevel(buf, gzip.BestCompression)
 	defer w.Close()
 	_, err := w.Write(b)
-	return buf.Bytes(), err
+	if err != nil {
+		return b, false, err
+	}
+	if buf.Len() > len(b) {
+		return b, false, err
+	}
+	c := make([]byte, buf.Len())
+	copy(c, buf.Bytes())
+	g.pool.Put(buf)
+	return c, true, err
 }
 
 func (g GzipCompression) StreamCompress(origin io.Reader) (io.Reader, error) {